refactor: separated with multiple source/header files.
